+++ src/main/java/repository/db/impl/DbReviewRepository.java
@@ -44,12 +44,12 @@ public class DbReviewRepository implements DbRepository<Review>{
             
             Timestamp timestamp = Timestamp.valueOf(review.getReviewDate());
             
-            statement.setInt(1, review.getReviewID());
+            statement.setLong(1, review.getReviewID());
             statement.setString(2, review.getReviewText());
             statement.setInt(3, review.getReviewScore());
             statement.setTimestamp(4, timestamp);
-            statement.setInt(5, review.getMovie().getMovieID());
-            statement.setInt(6, review.getUser().getUserID());
+            statement.setLong(5, review.getMovie().getMovieID());
+            statement.setLong(6, review.getUser().getUserID());
             
             statement.executeUpdate();
             
@@ -145,33 +145,33 @@ public class DbReviewRepository implements DbRepository<Review>{
             
             while(rs.next()) {
                 Review review = new Review();
-                review.setReviewID(rs.getInt("reviewID"));
+                review.setReviewID(rs.getLong("reviewID"));
                 review.setReviewText(rs.getString("reviewtext"));
                 review.setReviewScore(rs.getInt("reviewscore"));
                 review.setReviewDate(rs.getTimestamp("reviewdate").toLocalDateTime());
                 
                 Movie movie = new Movie();
-                movie.setMovieID(rs.getInt("movieID"));
+                movie.setMovieID(rs.getLong("movieID"));
                 movie.setName(rs.getString("name"));
                 movie.setReleaseDate(rs.getObject("releaseDate", LocalDate.class));
                 movie.setScore(Math.floor(rs.getDouble("score")* 100) / 100);
                 movie.setDescription(rs.getString("description"));
                 
                 Director director = new Director();
-                director.setDirectorID(rs.getInt("directorID"));
+                director.setDirectorID(rs.getLong("directorID"));
                 director.setFirstName(rs.getString("firstname"));
                 director.setLastName(rs.getString("lastname"));
                 director.setDateOfBirth(rs.getObject("dateofbirth", LocalDate.class));
                 
                 MoviePoster moviePoster = new MoviePoster();
-                moviePoster.setMoviePosterID(rs.getInt("movieposterID"));
+                moviePoster.setMoviePosterID(rs.getLong("movieposterID"));
                 moviePoster.setPosterImage(ImageIO.read(rs.getBlob("posterimage").getBinaryStream()));
                 
                 movie.setDirector(director);
                 movie.setMoviePoster(moviePoster);
                 
                 User user = new User();
-                user.setUserID(rs.getInt("userID"));
+                user.setUserID(rs.getLong("userID"));
                 user.setUsername(rs.getString("username"));
                 user.setPassword(rs.getString("password"));
                 user.setAdmin(rs.getBoolean("admin"));
@@ -206,33 +206,33 @@ public class DbReviewRepository implements DbRepository<Review>{
             
             while(rs.next()) {
                 Review review = new Review();
-                review.setReviewID(rs.getInt("reviewID"));
+                review.setReviewID(rs.getLong("reviewID"));
                 review.setReviewText(rs.getString("reviewtext"));
                 review.setReviewScore(rs.getInt("reviewscore"));
                 review.setReviewDate(rs.getTimestamp("reviewdate").toLocalDateTime());
                 
                 Movie movie = new Movie();
-                movie.setMovieID(rs.getInt("movieID"));
+                movie.setMovieID(rs.getLong("movieID"));
                 movie.setName(rs.getString("name"));
                 movie.setReleaseDate(rs.getObject("releaseDate", LocalDate.class));
                 movie.setScore(Math.floor(rs.getDouble("score")* 100) / 100);
                 movie.setDescription(rs.getString("description"));
                 
                 Director director = new Director();
-                director.setDirectorID(rs.getInt("directorID"));
+                director.setDirectorID(rs.getLong("directorID"));
                 director.setFirstName(rs.getString("firstname"));
                 director.setLastName(rs.getString("lastname"));
                 director.setDateOfBirth(rs.getObject("dateofbirth", LocalDate.class));
                 
                 MoviePoster moviePoster = new MoviePoster();
-                moviePoster.setMoviePosterID(rs.getInt("movieposterID"));
+                moviePoster.setMoviePosterID(rs.getLong("movieposterID"));
                 moviePoster.setPosterImage(ImageIO.read(rs.getBlob("posterimage").getBinaryStream()));
                 
                 movie.setDirector(director);
                 movie.setMoviePoster(moviePoster);
                 
                 User user = new User();
-                user.setUserID(rs.getInt("userID"));
+                user.setUserID(rs.getLong("userID"));
                 user.setUsername(rs.getString("username"));
                 user.setPassword(rs.getString("password"));
                 user.setAdmin(rs.getBoolean("admin"));
@@ -301,7 +301,7 @@ public class DbReviewRepository implements DbRepository<Review>{
     
     private Role loadRole(ResultSet rs) throws Exception{
         Actor actor=  new Actor();
-        actor.setActorID(rs.getInt("actorID"));
+        actor.setActorID(rs.getLong("actorID"));
         actor.setFirstName(rs.getString("firstname"));
         actor.setLastName(rs.getString("lastname"));
         actor.setBiography(rs.getString("biography"));
@@ -315,7 +315,7 @@ public class DbReviewRepository implements DbRepository<Review>{
     
     private MovieGenre loadMovieGenre(ResultSet rs) throws Exception{
         Genre genre = new Genre();
-        genre.setGenreID(rs.getInt("ggenreID"));
+        genre.setGenreID(rs.getLong("ggenreID"));
         genre.setName(rs.getString("gname"));
         
         MovieGenre movieGenre = new MovieGenre();
@@ -326,7 +326,7 @@ public class DbReviewRepository implements DbRepository<Review>{
     
     private Production loadProduction(ResultSet rs) throws Exception{
         ProductionCompany pc = new ProductionCompany();
-        pc.setProductionCompanyID(rs.getInt("pcID"));
+        pc.setProductionCompanyID(rs.getLong("pcID"));
         pc.setName(rs.getString("pcname"));
         
         Production production = new Production();
